Added animation dict
