+++ src/map/SemanticMarkers.tsx
@@ -2,11 +2,16 @@ import React, {cloneElement, ReactElement, useEffect, useState} from "react"; //
 import { GeoBBox, MarkerProps, BoundsViewport } from "./Types";
 import { useLeaflet } from "react-leaflet";
 import { LatLngBounds } from 'leaflet'
+import geohashAnimation from "./animation_functions/geohash";
 
 interface SemanticMarkersProps {
-  onViewportChanged: (bvp: BoundsViewport) => void,
+  onViewportChanged: (bvp: BoundsViewport, duration: number) => void,
   markers: Array<ReactElement<MarkerProps>>,
-  setMarkerElements: (markers: ReactElement<MarkerProps>[]) => void
+  animation: {
+    method: string,
+    duration: number,
+    // function: () => ReactElement<MarkerProps>[]
+  }
 }
 
 /**
@@ -15,7 +20,7 @@ interface SemanticMarkersProps {
  * 
  * @param props 
  */
-export default function SemanticMarkers({ onViewportChanged, markers, setMarkerElements}: SemanticMarkersProps) {
+export default function SemanticMarkers({ onViewportChanged, markers, animation}: SemanticMarkersProps) {
   const { map } = useLeaflet();
 
   const [prevMarkers, setPrevMarkers] = useState<ReactElement<MarkerProps>[]>(markers);
@@ -31,7 +36,7 @@ export default function SemanticMarkers({ onViewportChanged, markers, setMarkerE
       if (map != null) {
         const bounds = boundsToGeoBBox(map.getBounds());
         const zoomLevel = map.getZoom();
-        onViewportChanged({ bounds, zoomLevel });
+        onViewportChanged({ bounds, zoomLevel }, animation.duration);
       }
     }
 
@@ -45,39 +50,7 @@ export default function SemanticMarkers({ onViewportChanged, markers, setMarkerE
 
   useEffect(() => {
       if (markers.length > 0 && prevMarkers.length > 0) {
-          const prevGeoHash = prevMarkers[0].key as string;
-          const currentGeohash = markers[0].key as string;
-
-          /** Zoom Out - Move existing markers to new position
-           * Existing GeoHash = gcwr
-           * New Geohash      = gcw
-           **/
-          if (prevGeoHash.length > currentGeohash.length) {
-            setZoomType('out');
-            const hashDif = prevGeoHash.length - currentGeohash.length;
-            // Get a new array of existing markers with new position property
-            const cloneArray = updateMarkers(prevMarkers, markers, hashDif);
-            // Combine the new and existing markers
-            setConsolidatedMarkers([...markers, ...cloneArray]);
-          }
-          /** Zoom In - New markers start at old position
-           * Existing GeoHash = gcw
-           * New Geohash      = gcwr
-           **/
-          else if (prevGeoHash.length < currentGeohash.length) {
-            setZoomType('in');
-            const hashDif = currentGeohash.length - prevGeoHash.length;
-            // Get a new array of new markers with existing position property
-            const cloneArray = updateMarkers(markers, prevMarkers, hashDif);
-            // Set final render markers to the cloneArray which holds the new markers with
-            // their new starting location
-            setConsolidatedMarkers(cloneArray)
-          }
-          /** No difference in geohashes - Render markers as they are **/
-          else {
-            setZoomType(null);
-            setConsolidatedMarkers([...markers])
-          }
+        geohashAnimation({prevMarkers, markers, setZoomType, setConsolidatedMarkers})
       }
       /** First render of markers **/
       else {
@@ -149,24 +122,3 @@ function boundsToGeoBBox(bounds : LatLngBounds) : GeoBBox {
 	   northEast: [north, east] }
 }
 
-function updateMarkers(toChangeMarkers: Array<ReactElement<MarkerProps>>,
-                       sourceMarkers: Array<ReactElement<MarkerProps>>,
-                       hashDif: number) {
-  return toChangeMarkers.map((markerObj) => {
-    // Calculate the matching geohash
-    const sourceKey = markerObj.key as string;
-    const sourceHash = sourceKey.slice(0, -hashDif);
-
-    // Find the object with the matching geohash
-    const matchingMarkers = sourceMarkers.filter(obj => {
-      return obj.key === sourceHash
-    });
-
-    // Clone marker element with new position
-    const markerCloneProps = {
-      position: matchingMarkers[0].props.position
-    };
-    return cloneElement(markerObj, markerCloneProps);
-  });
-
-}
